+++ vrc-get-gui/src/main.rs
@@ -23,11 +23,8 @@ mod mcp;
 mod mcp_client_config;
 mod storage;
 mod templates;
-#[cfg(windows)]
 mod unity_process;
 
-#[cfg_attr(windows, path = "os_windows.rs")]
-#[cfg_attr(not(windows), path = "os_posix.rs")]
 mod os;
 mod state;
 mod updater;